refactor(rendering): extract 4k pipeline types
